Prevent windows from being left set to abandoned frames after xrandr
'group-remove-head' would use 'head-window' to lookup which windows
belonged to frames on the removed head, in order to reassign them to a
frame that should exist. This adds a 'frame-head' method for
'tile-groups' that will remain consistent will stump's internal state,
even if a window was physically relocated by X11 (such as during a
head removal).

Also various checks/fixups for if similar situations reoccur