Fix the ioloop timeout
1451b05 fixed the issues with unix-fast-select returning nil, however,
it replace a call to `multiple-value-call' to one with
`multiple-value-bind' to get the errno. However, (compute-timeouts)
returns multiple values: seconds and milliseconds remanining in the
timeout. 1451b05 passed the seconds value, and set the milliseconds to
nil, resulting in waiting for 0 seconds and 0 milliseconds whenever
the timeout was supposed to be less than a second

I also made the syscall-level errors from this debug level 1 instead
of 5