Fix crash when switching screens with floating groups
Fixes #1041.

If the number of screens drops to zero (e.g. when turning off the
monitor, or when unplugging an external screen from a laptop before
re-enabling a disabled built-in display) and floating groups exist,
the window manager will crash.  This is a regression introduced in
7fa64230 - a very helpful new generic function was added to preserve
the layout of tiling groups when switching screens, but no default
implementation was given for floating groups.