Fix crash when autoenabling minor mode thats already enabled
Currently when creating a new object which should have minor modes enabled in it
there comes an error of type minor-mode-autoenable-error. This patch addresses
this by A) establishing a restart around the autoenable error signalled in the
autoenable-minor-mode method to ignore the error, B) binding a handler to invoke
this restart in the replace-class around method specializing on (t t), and C)
explicitly checking if an object already has a minor mode enabled in it when
syncing minor modes for an object.