Print functions macros and commands in semi-pretty way in the manual
This commit adds a function to the manual.lisp file which prints a lambda list
stripped of any default values for optional and keyword arguments.  These are
then pretty printed below within a texinfo verbatim environment.  This is
intended to be used for functions macros and commands when generating the
manual, to avoid the issue of hard to read documentation (see issue #1055).  The
default values are pretty-printed in order A) keep the line length down, and B)
not print structure or literals.  Line length is kept to 80.  If default values
will have less than 30 characters to print within (due to the length of an
argument name), then they are printed on a new line.

Example:

(defun foo (a b &optional (c 'hi cpp) &key dee (e 'bye))
  "foo docstring"
  nil)

gets printed in the manual as

-- Function: foo a b &optional c &key dee e
    Default Values:
      c   'hi
      dee nil
      e   'bye
    foo docstring