Accept root-window button-presses when received for a source-window (child) that is ours.
In d5230923, mouse-grabbing was changed to bind to the actual managed
window instead of the parent window (cf. that commit for the details),
but this caused all button-presses on window decorations to become
ignored, and float groups to effectively stop working (at least as far
as moving and resizing are concerned).

This is a workaround for that problem that receives the button-press
when it is propagated to the root-window instead, as indicated by the
event child-window being our decoration window, translating the
coordinates as required.