Patch to redo format-expand
There was a fixme saying the previous implementation was horribly inefficient,
so I tried to improve it. The replacement is, um, a little bigger. But. It's
technically more efficient?

Also, defined a function - replace-ranges - that may be useful elsewhere, maybe.

But, this format-expand doesn't actually run that much quicker in some shorter cases. In complicated
cases, yes, and it scales better, but the main improvement is that it almost
always conses a lot less? From only consing 70% as much in short cases, to
consing less than 10% as much in more complicated ones. (And by complicated I mean both
having longer fmt strings, and those fmt strings having a larger quantity of
expansions.)

Though, the consing it did was when I was running it in loops of 1000000, with
all expanders just (constantly "some-string") with no args, since the performance of the
expanders themselves isn't what's being tested. They both access and call them
in the same way, after all.

I tried to make format-expand produce output that matched the old version as
consistently as possible, but, I might have missed things! So.