Patch for mode-line functions alternate and scroll
There was a fixme that wanted multiple instances of each to be able to coexist, so.

This does two things - one, provides make-string-alternator and
make-string-scroller, which each return a lambda that can be passed the args of
their old functions, they just have an internal state (or, well. External?
Lexical binds). So, store that and funcall it.

Also, they store the last string they returned and only update their return if
the period has passed, otherwise returning the stored value.

This also defines the macros `alternate` and `scroll` that kind of
work as drop-ins for the old functions, except where macros don't work. Like funcalls/apply's.

And switched the bar function from using format to making a string and filling
it. It's a bit bigger, but format wasn't really needed? Maybe?