input(input-refine-fuzzy): Add for fuzzy matching.
This way, one can filter through the suggestions in a way that's
slightly better than prefix, but less powerful than regexp one.

The implementation is somewhat simplistic. In particular, it doesn't
check the order of space-separated parts and doesn't ensure
that all the matches are unique. Still, better than prefix matching.