remove number zero from group number map
StumpWM expects group numbers to grow incrementally and for their numbers to
match the *group-number-map*. However because group numbers start from one the
tenth group created gets the group number ten, while the *group-number-map* and
associated function group-map-number report its number as zero, because zero is
the tenth character in that string. Removing the zero from the string removes
this inconsistency, as any group-map-number that isnt present in the string is
returned via princ-to-string.