Patch MARK to accept the window as an optional argument
Modify the MARK command to take two optional arguments. Current behavior is
preserved. The arguments are the window to mark, which defaults to the current
window, and a new argument controlling whether or not to display a message to
the user that the window has been marked. In addition, the message displayed has
been modified to also display the window that has been marked.