Use non-canonical paths when resolving default data dir
Resolving the data dir path based on probe-file's canonical path lookup of pathspec breaks users that manage their StumpWM configuration through symlinked dotfiles. The data dir would incorrectly be set to wherever the root of the symlinked file is located on the file system. This may be a read only location on e.g. systems such as Guix and NixOS, or just some random location in the case of stow. StumpWM would then incorrectly treat this random location as its data directory. This patch ensures we operate on the pathspec argument and not the truename result from probe-file.