Make find-mode-line-by-window find mode-lines with matching xwin ids
See https://github.com/polybar/polybar/issues/1853

The :destroy-notify handler is meant to destroy mode-line windows if
the event applies to one. However, find-mode-line-by-window searched
for a window object matching with eql rather than a window id. Since
the event window would not eql the mode-line-window, find returned nil
regardless of whether the window ids match, meaning that the mode-line
window would not be destroyed.

This change uses window id as the key for find and searches for the id
of the given window.